Use python from the system
i.e., do use use a pre-built version downloaded from the internet.

This is a workaround to avoid the default behavior of bazel/envoy that
is to use a hermetic python version. Just like other toolchain (e.g. go,
java, etc) we should rely on the system version.

This works by modifying BUILD.bazel that is installed automatically by
bazel to make it use the system's python. It turns out it is a small
change. Then we modify `update-deps.sh` script to not vendor python and
instead vendor this slightly modified BUILD.bazel.